auth en nominas y detalles

+++ controllers/recursosHumanos/detallesEmpleados.controller.js
@@ -1,9 +1,15 @@
 import { Op } from "sequelize";
 import rh from "../../models";
+import auth from "../../models/auth.model";
+const { verificar } = auth;
 const { departamentos, tiposNominas, empleados, detalleEmpleado } = rh;
+import { guardarBitacora } from "../../models/auditorias";
+const area = "Referencias Bancarias";
 
 export async function buscarDetalles(req, res) {
   try {
+    let user = verificar(req.headers.authorization);
+    if (!user.success) throw user;
     const response = await detalleEmpleado.findAll({
       include: [
         {
@@ -17,6 +23,7 @@ export async function buscarDetalles(req, res) {
     });
     res.status(200).json({ success: true, response });
   } catch (err) {
+    console.log(err);
     res
       .status(400)
       .json({ success: false, err, msg: "Error al obtener la nomina" });
@@ -25,6 +32,8 @@ export async function buscarDetalles(req, res) {
 
 export async function buscarDetallesXEmpleados(req, res) {
   try {
+    let user = verificar(req.headers.authorization);
+    if (!user.success) throw user;
     const response = [];
     const empleadosD = await empleados.findAll({
       where: {
@@ -53,6 +62,8 @@ export async function buscarDetallesXEmpleados(req, res) {
 
 export async function crearDetalles(req, res) {
   try {
+    let user = verificar(req.headers.authorization);
+    if (!user.success) throw user;
     const { idEmpleado, idTipoNomina } = req.body;
 
     const validar = await detalleEmpleado.findOne({
@@ -65,8 +76,15 @@ export async function crearDetalles(req, res) {
       idempleado: idEmpleado,
       idtipo_nomina: idTipoNomina,
     });
+    await guardarBitacora([
+      area,
+      user.token.data.datos.idempleado,
+      2,
+      idEmpleado,
+    ]);
     res.status(200).json({ success: true, response });
   } catch (err) {
+    console.log(err);
     res
       .status(400)
       .json({ success: false, err, msg: "Error al guardar la nomina" });
@@ -75,6 +93,8 @@ export async function crearDetalles(req, res) {
 
 export async function editarDetalle(req, res) {
   try {
+    let user = verificar(req.headers.authorization);
+    if (!user.success) throw user;
     const { idEmpleado } = req.params;
     const { idTipoNomina } = req.body;
 
@@ -84,9 +104,16 @@ export async function editarDetalle(req, res) {
       },
       { where: { idempleado: idEmpleado } }
     );
-
+    await guardarBitacora([
+      area,
+      user.token.data.datos.idempleado,
+      3,
+      idEmpleado,
+    ]);
     res.status(200).json({ success: true, response });
   } catch (err) {
+    console.log(err);
+
     res
       .status(400)
       .json({ success: false, err, msg: "Error al guardar la nomina" });
@@ -95,12 +122,20 @@ export async function editarDetalle(req, res) {
 
 export async function eliminarDetalle(req, res) {
   try {
+    let user = verificar(req.headers.authorization);
+    if (!user.success) throw user;
     const { idEmpleado } = req.params;
 
     const response = await detalleEmpleado.destroy({
       where: { idempleado: idEmpleado },
     });
 
+    await guardarBitacora([
+      area,
+      user.token.data.datos.idempleado,
+      4,
+      idEmpleado,
+    ]);
     res.status(200).json({ success: true, response });
   } catch (err) {
     res
