+++ controllers/recursosHumanos/nominas.controller.js
@@ -1,10 +1,17 @@
 import { Op } from "sequelize";
 import rh from "../../models";
 import departamento from "../../models/recursosHumanos/departamentos.model";
+import { guardarBitacora } from "../../models/auditorias";
+import auth from "../../models/auth.model";
 const { nominas, tiposNominas, empleados } = rh;
+const { verificar } = auth;
+const area = "Nominas";
+const areaSub = "Tipo Nomina";
 
 export async function guardarNomina(req, res) {
   try {
+    let user = verificar(req.headers.authorization);
+    if (!user.success) throw user;
     const { idEmpleado, fecha, monto, idTipoNomina } = req.body;
     const response = await nominas.create({
       idempleado: idEmpleado,
@@ -12,6 +19,13 @@ export async function guardarNomina(req, res) {
       idtipo_nomina: idTipoNomina,
       monto,
     });
+    await guardarBitacora([
+      area,
+      user.token.data.datos.idempleado,
+      2,
+      response.idturno,
+    ]);
+
     res.status(200).json({ success: true, response });
   } catch (err) {
     res
@@ -22,6 +36,8 @@ export async function guardarNomina(req, res) {
 
 export async function obtenerNominas(req, res) {
   try {
+    let user = verificar(req.headers.authorization);
+    if (!user.success) throw user;
     const { fechaI, fechaF } = req.query;
     const response = await nominas.findAll({
       include: [
@@ -42,6 +58,8 @@ export async function obtenerNominas(req, res) {
 
 export async function obtenerUltimoRegistro(req, res) {
   try {
+    let user = verificar(req.headers.authorization);
+    if (!user.success) throw user;
     const response = await nominas.findOne({
       include: [
         {
@@ -60,6 +78,8 @@ export async function obtenerUltimoRegistro(req, res) {
 
 export async function obtenerNomina(req, res) {
   try {
+    let user = verificar(req.headers.authorization);
+    if (!user.success) throw user;
     const { idNomina } = req.params;
     const response = await nominas.findOne({
       include: [
@@ -80,6 +100,8 @@ export async function obtenerNomina(req, res) {
 
 export async function actualizarNomina(req, res) {
   try {
+    let user = verificar(req.headers.authorization);
+    if (!user.success) throw user;
     const { idNomina } = req.params;
     const { idEmpleado, fecha, monto, idTipoNomina } = req.body;
     const cuerpo = {
@@ -91,6 +113,12 @@ export async function actualizarNomina(req, res) {
     const response = await nominas.update(cuerpo, {
       where: { idnomina: idNomina },
     });
+    await guardarBitacora([
+      area,
+      user.token.data.datos.idempleado,
+      3,
+      idNomina,
+    ]);
     res.status(200).json({ success: true, response });
   } catch (err) {
     res.status(400).json({ success: false, err });
@@ -99,10 +127,18 @@ export async function actualizarNomina(req, res) {
 
 export async function eliminarNomina(req, res) {
   try {
+    let user = verificar(req.headers.authorization);
+    if (!user.success) throw user;
     const { idNomina } = req.params;
     const response = await nominas.destroy({
       where: { idnomina: idNomina },
     });
+    await guardarBitacora([
+      area,
+      user.token.data.datos.idempleado,
+      4,
+      idNomina,
+    ]);
     res.status(200).json({ success: true, response });
   } catch (err) {
     res.status(400).json({ success: false, err });
@@ -111,11 +147,19 @@ export async function eliminarNomina(req, res) {
 
 export async function guardarTipoNomina(req, res) {
   try {
+    let user = verificar(req.headers.authorization);
+    if (!user.success) throw user;
     const { tipo, banco } = req.body;
     const response = await tiposNominas.create({
       tipo,
       banco,
     });
+    await guardarBitacora([
+      areaSub,
+      user.token.data.datos.idempleado,
+      2,
+      response.idtipo_nomina,
+    ]);
     res.status(200).json({ success: true, response });
   } catch (err) {
     res.status(400).json({ success: false, err: err, msg: "asd" });
@@ -124,6 +168,8 @@ export async function guardarTipoNomina(req, res) {
 
 export async function editarTipoNomina(req, res) {
   try {
+    let user = verificar(req.headers.authorization);
+    if (!user.success) throw user;
     const { idTipo } = req.params;
     const { tipo, banco } = req.body;
     const response = await tiposNominas.update(
@@ -133,6 +179,12 @@ export async function editarTipoNomina(req, res) {
       },
       { where: { idtipo_nomina: idTipo } }
     );
+    await guardarBitacora([
+      areaSub,
+      user.token.data.datos.idempleado,
+      3,
+      idTipo,
+    ]);
     res.status(200).json({ success: true, response });
   } catch (err) {
     res.status(400).json({ success: false, err: err, msg: "asd" });
@@ -141,10 +193,18 @@ export async function editarTipoNomina(req, res) {
 
 export async function eliminarTipoNomina(req, res) {
   try {
+    let user = verificar(req.headers.authorization);
+    if (!user.success) throw user;
     const { idTipo } = req.params;
     const response = await tiposNominas.destroy({
       where: { idtipo_nomina: idTipo },
     });
+    await guardarBitacora([
+      areaSub,
+      user.token.data.datos.idempleado,
+      4,
+      idTipo,
+    ]);
     res.status(200).json({ success: true, response });
   } catch (err) {
     res.status(400).json({ success: false, err: err, msg: "asd" });
@@ -153,11 +213,12 @@ export async function eliminarTipoNomina(req, res) {
 
 export async function obtenerTipoNomina(req, res) {
   try {
+    let user = verificar(req.headers.authorization);
+    if (!user.success) throw user;
     const response = await tiposNominas.findAll();
-    console.log({ response });
 
     res.status(200).json({ success: true, response });
   } catch (err) {
-    res.status(400).json({ success: false, err: err, msg: "asd" });
+    res.status(400).json({ success: false, err: err });
   }
 }
